backend/requirements.txt
flask-openapi3[swagger,redoc]==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
pymysql==1.1.0
markdown==3.6
pygments==2.18.0
python-dotenv==1.0.1
cryptography==42.0.8
Pillow==10.3.0
requests==2.32.3
